requirements.txt
asgiref==3.3.1
cairocffi==1.2.0
CairoSVG==2.5.0
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cssselect2==0.4.1
defusedxml==0.6.0
Django==3.1.3
django-countries==7.0
django-crispy-forms==1.10.0
django-environ==0.4.5
django-resized==0.3.11
django-taggit==1.3.0
django-weasyprint==1.0.2
djangorestframework==3.11.1
drf-yasg==1.20.0
html5lib==1.1
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
packaging==20.7
Pillow==8.0.1
psycopg2-binary==2.8.6
pycparser==2.20
pyparsing==2.4.7
Pyphen==0.10.0
pytz==2020.4
requests==2.25.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.15.0
sqlparse==0.4.1
tinycss2==1.1.0
uritemplate==3.0.1
urllib3==1.26.2
WeasyPrint==52.1
webencodings==0.5.1
